Add googletest and test cases. Fix bugs. (#1)
* Add gtest as a submodule

* Change project structure; add test cases.

* Bug fix in amax, amin, and aum.

* Add test cases and fix bugs.

+++ core/blas/src/level-one/asum.cu
@@ -1,5 +1,5 @@
-#include "../cncblas.h"
-#include "helpers.cuh"
+#include "cncblas.h"
+#include "src/helpers.cuh"
 #include <stdexcept>
 #include <iostream>
 
@@ -98,7 +98,7 @@ __global__ void cncblasDasumKernel(size_t n, const double *x, double *result) {
   sdata[tid] = 0;
   size_t i = idx;
   while (i < n) {
-    sdata[tid] += fabsf(x[i]);
+    sdata[tid] += fabs(x[i]);
     i += blockDim.x * gridDim.x;
   }
   __syncthreads();
